// x-pack/legacy/plugins/lens/public/map_plugin/types.ts
/*
 * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
 * or more contributor license agreements. Licensed under the Elastic License;
 * you may not use this file except in compliance with the Elastic License.
 */

export interface State {
    layerId: string;
    accessor: string;
  }
  
export type PersistableState = State;
  